+++ testdata/app/validators/admin/access_validator.rb
@@ -1,0 +1,14 @@
+module Admin
+  class AccessValidator
+    include ActiveModel::Validations
+
+    attr_reader :role
+
+    validates :role, presence: true
+    validates :role, inclusion: { in: %w[admin superadmin] }
+
+    def initialize(role)
+      @role = role
+    end
+  end
+end