+++ testdata/app/validators/concerns/rule_helpers.rb
@@ -1,0 +1,5 @@
+module RuleHelpers
+  def blank_or_matches?(value, format)
+    value.blank? || value.match?(format)
+  end
+end